+++ front-end/legeno-around-here/src/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import {BrowserRouter, Redirect, Route, Switch} from 'react-router-dom';
-import Join from './components/pages/Join';
-import Login from './components/pages/Login';
+import Join from './components/pages/JoinPage';
+import Login from './components/pages/LoginPage';
 import PostingPage from './components/pages/posting/PostingPage';
 import MyProfilePage from './components/pages/myProfile/MyProfilePage';
 import SectorPage from './components/pages/sector/SectorPage';
 import PostDetailPage from './components/pages/post/PostDetailPage';
-import Home from './components/pages/Home';
-import Ranking from './components/pages/Ranking';
+import HomePage from './components/pages/HomePage';
+import RankingPage from './components/pages/RankingPage';
 
 function App() {
   const mainArea = localStorage.getItem('mainAreaName');
@@ -20,15 +20,15 @@ function App() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route path="/" exact component={Home} />
+        <Route path="/" exact component={HomePage} />
         <Route path="/join" exact component={Join} />
         <Route path="/login" exact component={Login} />
         <Route path="/myProfile" exact component={MyProfilePage} />
         <Route path="/posting" exact component={PostingPage} />
         <Route path="/sector" exact component={SectorPage} />
         <Route path="/posts/:postId" exact component={PostDetailPage} />
-        <Route path="/home" exact component={Home} />
-        <Route path="/ranking" exact component={Ranking} />
+        <Route path="/home" exact component={HomePage} />
+        <Route path="/ranking" exact component={RankingPage} />
         <Redirect path="*" to="/" />
       </Switch>
     </BrowserRouter>
