+++ front-end/legeno-around-here/src/components/pages/JoinPage.js
